// api/types/PageNewsResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { NewsResponse } from "./NewsResponse.ts";
import type { PageableObject } from "./PageableObject.ts";
import type { SortObject } from "./SortObject.ts";

export type PageNewsResponse = {
    /**
     * @type integer | undefined, int64
    */
    totalElements?: number;
    /**
     * @type integer | undefined, int32
    */
    totalPages?: number;
    /**
     * @type boolean | undefined
    */
    first?: boolean;
    /**
     * @type boolean | undefined
    */
    last?: boolean;
    /**
     * @type integer | undefined, int32
    */
    size?: number;
    /**
     * @type array | undefined
    */
    content?: NewsResponse[];
    /**
     * @type integer | undefined, int32
    */
    number?: number;
    /**
     * @type object | undefined
    */
    sort?: SortObject;
    /**
     * @type integer | undefined, int32
    */
    numberOfElements?: number;
    /**
     * @type object | undefined
    */
    pageable?: PageableObject;
    /**
     * @type boolean | undefined
    */
    empty?: boolean;
};